ai-ml/requirements.txt
# AI/ML Module Requirements for SmartCitySense - Member B2
# Vision & Predictive Modeling Dependencies

# ========================================
# CORE ML/AI FRAMEWORKS
# ========================================
torch>=2.0.0                    # PyTorch for deep learning
torchvision>=0.15.0             # Computer vision models
ultralytics>=8.0.0              # YOLOv8 for object detection
opencv-python>=4.8.0            # Image/video processing
opencv-python-headless>=4.8.0   # Headless CV for servers
pillow>=10.0.0                  # Image manipulation
numpy>=1.24.0                   # Numerical computations
scikit-learn>=1.3.0             # ML algorithms (IsolationForest)
scikit-image>=0.21.0            # Advanced image processing

# ========================================
# TIME SERIES & FORECASTING
# ========================================
prophet>=1.1.5                  # Facebook Prophet for forecasting
statsmodels>=0.14.0             # Statistical modeling
pandas>=2.0.0                   # Data manipulation
scipy>=1.11.0                   # Scientific computing

# ========================================
# COMPUTER VISION ENHANCEMENTS
# ========================================
transformers>=4.30.0            # Hugging Face models (CLIP, BLIP, BERT)
timm>=0.9.0                     # PyTorch Image Models

# ========================================
# TEXT PROCESSING & NLP (Member B1)
# ========================================
langchain>=0.1.0                # LLM orchestration framework
langchain-google-genai>=0.0.5   # Gemini integration for langchain
langchain-openai>=0.0.5         # OpenAI integration for langchain
google-generativeai>=0.3.0      # Google Gemini API
openai>=1.0.0                   # OpenAI GPT API
sentencepiece>=0.1.99           # Tokenization for multilingual models
sentence-transformers>=2.2.0    # Semantic embeddings
textblob>=0.17.0                # Simple NLP (fallback)
nltk>=3.8                       # Natural Language Toolkit

# ========================================
# API & WEB FRAMEWORK
# ========================================
fastapi>=0.104.0                # REST API framework
uvicorn[standard]>=0.24.0       # ASGI server
pydantic>=2.0.0                 # Data validation
python-multipart>=0.0.6         # File upload support
aiofiles>=23.0.0                # Async file operations

# ========================================
# FIREBASE & DATABASE
# ========================================
firebase-admin>=6.2.0           # Firebase integration
google-cloud-firestore>=2.11.0  # Firestore client

# ========================================
# UTILITIES
# ========================================
python-dotenv>=1.0.0            # Environment variables
requests>=2.31.0                # HTTP client
httpx>=0.25.0                   # Async HTTP client
python-json-logger>=2.0.7       # Structured logging

# ========================================
# TESTING
# ========================================
pytest>=7.4.0                   # Testing framework
pytest-asyncio>=0.21.0          # Async test support
pytest-cov>=4.1.0               # Code coverage
httpx>=0.25.0                   # API testing

# ========================================
# MONITORING & LOGGING
# ========================================
loguru>=0.7.0                   # Advanced logging
psutil>=5.9.0                   # System monitoring
tqdm>=4.66.0                    # Progress bars

# ========================================
# DATA PROCESSING
# ========================================
matplotlib>=3.7.0               # Plotting (for analysis)
seaborn>=0.12.0                 # Statistical visualization
plotly>=5.17.0                  # Interactive plots

# ========================================
# OPTIONAL: GPU ACCELERATION
# ========================================
# For CUDA-enabled systems, install PyTorch with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# ========================================
# DEVELOPMENT TOOLS
# ========================================
black>=23.0.0                   # Code formatter
flake8>=6.0.0                   # Linter
mypy>=1.5.0                     # Type checking
ipython>=8.14.0                 # Enhanced REPL
jupyter>=1.0.0                  # Notebooks for experimentation
